Move pages to the new screens folder
